requirements.txt
fastapi
pymongo
python-dotenv
uvicorn
pytz
datetime
typing
logging
PyPDF2
python-multipart